Initial Makefile workflow
I have redone some of the initial steps, now using a Makefile along with
simple scripts. So far it downloads the necessary data, does some data
formatting, and does a 2D scatter plot and contour plot using the basic
data.

This new workflow makes it easy for others to reproduce the results, and
saves computation time when making changes to one portion of the code
and re-running.