Incomplete Makefile workflow code
Over the summer, in my spare time I rewrote my previous code mostly from scratch, this time orchestrating everything from a Makefile, which runs a number of small scripts.

The distance estimates are questionable, so there is probably a bug somewhere. The nice thing about this workflow is that it reproduces all the results with a single run of `make all` and if anything changes, only the results affected by those changes are re-run.